refactor(cip2): use BigIntMath from core package
docs(cip2): add usage example in README, fix a few comments

chore(cip2): rm trash code

refactor(cip2): split the algorithm into multiple files

feat(cip2): prioritize picking utxo with ada-only when assets are not needed

fix(cip2): fix coalescing change values for min ada requirement

refactor(cip2): convert some loops to for-in

chore(cip2): add core package to package.json dependencies, cleanup some comments
