+++ packages/cip2/package.json
@@ -27,6 +27,7 @@
   },
   "dependencies": {
     "@cardano-sdk/cardano-serialization-lib": "0.1.0",
+    "@cardano-sdk/core": "0.1.0",
     "lodash-es": "^4.17.21"
   }
 }
